2012-06-09 version 1.4.3 	Added Python script binding for Python 2.7.3. (warning -- may memory leak) 	Function fromVariant -- now integer 0 can be auto casted to NULL pointer. 	Fixed a bug that the ownerItem in GMetaItem in a dangling define may point to bad pointer.
git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@400 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ src/gmetaclass.cpp
@@ -602,7 +602,7 @@ void GMetaClass::extractTo(GMetaClass * master)
 			this->implement->itemLists[c]->extractTo(master->implement->itemLists[c]);
 		}
 	}
-	for(int i = 0; i < this->implement->metaList.getCount(); ++i) {
+	for(size_t i = 0; i < this->implement->metaList.getCount(); ++i) {
 		GMetaItem * item = this->implement->metaList.getItemAt(i);
 		if(item != NULL) {
 			item->ownerItem = master;
